+++ ru.taximaxim.codekeeper.core.tests/src/main/resources/ru/taximaxim/codekeeper/core/depcies/add_func_with_regcast_usr_func_regoperator_new.sql
@@ -1,0 +1,3 @@
+CREATE OR REPLACE FUNCTION public.f_regoperator(x integer) RETURNS text
+    LANGUAGE sql
+    AS $$ select 1; SELECT ('public.||==(text, text)'::regoperator)::text AS text;$$;